move implementation of H5CError class to its header file to avoid C4275 warning
